+++ changelogs/1.8.0.md
@@ -1,0 +1,3 @@
+**Misc**
+
+- Migrate to quilt mappings
