+++ common/src/main/resources/debugify-common.mixins.json
@@ -19,7 +19,9 @@
     "client.mc140646.TextFieldWidgetMixin",
     "client.mc145748.SliderWidgetMixin",
     "client.mc145929.InGameHudMixin",
+    "client.mc147605.AbstractParentElementMixin",
     "client.mc147605.ClickableWidgetMixin",
+    "client.mc147605.TextFieldWidgetMixin",
     "client.mc148149.OperatingSystemMixin",
     "client.mc151412.AddServerScreenMixin",
     "client.mc159163.ClientPlayNetworkHandlerMixin",
@@ -35,6 +37,7 @@
     "client.mc234898.RealmsMainScreenMixin",
     "client.mc249021.RealmsMainScreenMixin",
     "client.mc249059.DownloadingTerrainScreenMixin",
+    "client.mc4490.FishingBobberEntityRendererMixin",
     "client.mc46766.MinecraftClientMixin",
     "client.mc53312.VillagerResemblingModelMixin",
     "client.mc79545.InGameHudMixin",
@@ -42,8 +45,6 @@
     "client.mc93384.LivingEntityMixin"
   ],
   "mixins": [
-    "client.mc147605.AbstractParentElementMixin",
-    "client.mc147605.TextFieldWidgetMixin",
     "server.mc100991.FishingBobberEntityMixin",
     "server.mc100991.ProjectileEntityMixin",
     "server.mc119417.ServerPlayerEntityMixin",
