+++ forge/build.gradle.kts
@@ -94,10 +94,6 @@ components["java"].withGroovyBuilder {
     }
 }
 
-quiltflower {
-    addToRuntimeClasspath.set(true)
-}
-
 val minecraftVersion: String by rootProject
 
 modrinth {
