+++ gradle.properties
@@ -1,8 +1,6 @@
 org.gradle.jvmargs=-Xmx2G
 
-# https://fabricmc.net/develop
 minecraftVersion=1.18.2
-yarnVersion=1.18.2+build.+
 fabricLoaderVersion=0.13.+
 forgeVersion=40.0.+
